requirements.txt
python-dotenv
pinecone-client
langchain-openai
langchain-community
langchain-pinecone
langchain-text-splitters